fold scraper block into markdown-negotiation (single edge fn, ordering-proof)
The separate bot-mitigation function created two edge functions on /*, whose
execution order is not guaranteed - if markdown-negotiation ran first on a page
it would context.next() (fetch HTML = bandwidth) before the block fired. And
markdown-negotiation excludedPath's images/.json, so only bot-mitigation could
cover /search.json (136 GB) and images.

Fold the isScraper check into markdown-negotiation as its first line: one
function, block runs before any origin fetch, no ordering dependency, and one
fewer invocation per page. Widen scope to also cover images and .json (removed
from excludedPath) so the block reaches search.json and images; keep excluding
cheap hashed/cacheable assets (js/css/maps/fonts/stylesheets/md/txt/xml).
Delete bot-mitigation.ts. isScraper unit-tested; bundles clean with esbuild.
